requirements.txt
fastapi
pydantic
uvicorn
ollama
python-multipart
Pillow
structlog
python-dotenv
groq
openai